+++ loginradius-sdk/sdk/api/social/socialApi.js
@@ -30,12 +30,14 @@ module.exports = function (config) {
   /**
   * The Refresh Access Token API is used to refresh the provider access token after authentication. It will be valid for up to 60 days on LoginRadius depending on the provider. In order to use the access token in other APIs, always refresh the token using this API.<br><br><b>Supported Providers :</b> Facebook,Yahoo,Google,Twitter, Linkedin.<br><br> Contact LoginRadius support team to enable this API.
   * @param {accessToken} Uniquely generated identifier key by LoginRadius that is activated after successful authentication.
-  * @param {expiresIn} Allows you to specify a desired expiration time in minutes for the newly issued access_token.
+  * @param {expiresIn} Allows you to specify a desired expiration time in minutes for the newly issued access token.
+  * @param {isWeb} Is web or not.
   * @return Response containing Definition of Complete Token data
   *20.2
   */
 
-  module.refreshAccessToken = function (accessToken, expiresIn) {
+  module.refreshAccessToken = function (accessToken, expiresIn,
+    isWeb) {
     if (helper.isNullOrWhiteSpace(accessToken)) {
       return Promise.reject(helper.getValidationMessage('accessToken'));
     }
@@ -45,6 +47,9 @@ module.exports = function (config) {
     queryParameters.secret = config.apiSecret;
     if (expiresIn !== null) {
       queryParameters.expiresIn = expiresIn;
+    }
+    if (isWeb !== null) {
+      queryParameters.isWeb = isWeb;
     }
 
     var resourcePath = 'api/v2/access_token/refresh';